feat: move track view port to hook
